requirements.txt
pandas
nltk
gensim
scikit-learn
google-api-python-client
python-dotenv
emoji
numpy
fastapi
uvicorn
uvicorn[standard]
